+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
  - `FillValueMetadata::try_as_uint/int` can both now handle `FillValueMetadata::Uint/Int`
 
+### Fixed
+ - `Array::store_chunk` now erases empty chunks
+
 ## [0.5.0] - 2023-10-08
 
 ### Added