+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
  - `Array::store_chunk` now erases empty chunks
+ - Fixed a race in `Array::store_chunk_subset` and add a fast path if the subset spans the whole chunk
 
 ## [0.5.0] - 2023-10-08